pyproject.toml
[tool.poetry]
name = "ore-oh"
version = "0.1.0"
description = ""
authors = ["d3adcode <d3adcode@null.net>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
gunicorn = "^20.0.4"
ariadne = "^0.12.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
